% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FederatedDualAveraging.R
\name{clientUpdateDA}
\alias{clientUpdateDA}
\title{One client's K-step dual-averaging loop}
\usage{
clientUpdateDA(clientData, serverBroadcast, config)
}
\arguments{
\item{clientData}{local client data list with \code{xMatrix} and \code{yLabels}}

\item{serverBroadcast}{server state broadcast to the client}

\item{config}{algorithm configuration list}
}
\value{
numeric deltaZ = z_final - z_initial
}
\description{
One client's K-step dual-averaging loop
}
