% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clientUpdateDualAveragingCpp}
\alias{clientUpdateDualAveragingCpp}
\title{Run a C++ dual-averaging client update}
\usage{
clientUpdateDualAveragingCpp(clientData, serverBroadcast, config)
}
\arguments{
\item{clientData}{local client data list}

\item{serverBroadcast}{server state broadcast to the client}

\item{config}{algorithm configuration list}
}
\value{
client update report
}
\description{
Run a C++ dual-averaging client update
}
