% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{clusterClearState}
\alias{clusterClearState}
\title{Clear FederatedLearning state from cluster workers}
\usage{
clusterClearState(cl)
}
\arguments{
\item{cl}{cluster object}
}
\value{
invisibly, one NULL per worker
}
\description{
Clear FederatedLearning state from cluster workers
}
