% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{clusterCreateMatrices}
\alias{clusterCreateMatrices}
\title{Create local model matrices on each cluster worker}
\usage{
clusterCreateMatrices(cl, config)
}
\arguments{
\item{cl}{cluster object}

\item{config}{list containing \code{mapping} and model matrix options}
}
\value{
list of NULL values, one per worker
}
\description{
Create local model matrices on each cluster worker
}
