% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{clusterEvaluateModel}
\alias{clusterEvaluateModel}
\title{Evaluate model weights on every client with multiple metrics}
\usage{
clusterEvaluateModel(cl, w, threshold = 1e-04)
}
\arguments{
\item{cl}{cluster object}

\item{w}{numeric model coefficients}

\item{threshold}{density threshold}
}
\value{
data.frame with AUC, log loss, calibration intercept/slope
}
\description{
Evaluate model weights on every client with multiple metrics
}
