% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{clusterInit}
\alias{clusterInit}
\title{Initialize a worker cluster for client execution}
\usage{
clusterInit(clientHosts, clientPaths, mirai = TRUE)
}
\arguments{
\item{clientHosts}{worker host names or addresses}

\item{clientPaths}{client data paths; used to define cluster size}

\item{mirai}{if TRUE, use \code{mirai}; otherwise use PSOCK workers}
}
\value{
cluster object
}
\description{
Initialize a worker cluster for client execution
}
