% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{clusterLoadData}
\alias{clusterLoadData}
\title{Load PLP data on each cluster worker}
\usage{
clusterLoadData(cl, clientPaths, popSettings)
}
\arguments{
\item{cl}{cluster object}

\item{clientPaths}{paths to client PLP data folders}

\item{popSettings}{PatientLevelPrediction population settings}
}
\value{
integer vector of per-client population sizes
}
\description{
Load PLP data on each cluster worker
}
