% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep.R
\name{createClientMatrix}
\alias{createClientMatrix}
\title{Create a local sparse design matrix from PLP data and a global map}
\usage{
createClientMatrix(plpData, config)
}
\arguments{
\item{plpData}{a PLP data object with a populated \code{population}}

\item{config}{list containing \code{mapping} and \code{intercept}}
}
\value{
list with \code{xMatrix}, summary vectors, labels, and row count
}
\description{
Create a local sparse design matrix from PLP data and a global map
}
