% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep.R
\name{createGlobalMap}
\alias{createGlobalMap}
\title{Create a global covariateId → columnId mapping}
\usage{
createGlobalMap(covRefList, type = c("union", "intersection"))
}
\arguments{
\item{covRefList}{a list of data.frames, each with a column \code{covariateId}}

\item{type}{"union" or "intersection"}
}
\value{
data.frame with covariateId and columnId (1..P_global)
}
\description{
Create a global covariateId → columnId mapping
}
