% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FederatedCv.R
\name{evaluateClient}
\alias{evaluateClient}
\title{Evaluate model weights on a single client}
\usage{
evaluateClient(cl, w, config)
}
\arguments{
\item{cl}{cluster object}

\item{w}{numeric vector of model coefficients}

\item{config}{model configuration list}
}
\value{
named list of metrics: accuracy, auc, logloss, density
}
\description{
Evaluate model weights on a single client
}
