% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FederatedLearning.R
\docType{package}
\name{FederatedLearning}
\alias{FederatedLearning-package}
\alias{FederatedLearning}
\title{FederatedLearning}
\description{
A package for running federated learning algorithms in R.
}
\author{
\strong{Maintainer}: Egill Fridgeirsson \email{e.fridgeirsson@erasmusmc.nl}

}
\keyword{internal}
