% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FederatedCv.R
\name{federatedNestedCv}
\alias{federatedNestedCv}
\title{Outer + inner CV for federated hyperparameter tuning}
\usage{
federatedNestedCv(
  clientHosts,
  clientPaths,
  popSettings,
  algorithm,
  hyperGrid,
  rounds,
  clientFrac,
  resultDirectory,
  epsilon = 1e-06,
  mirai = FALSE
)
}
\arguments{
\item{clientHosts}{worker host names or addresses}

\item{clientPaths}{paths to client PLP data folders}

\item{popSettings}{PatientLevelPrediction population settings}

\item{algorithm}{registered federated learning algorithm name}

\item{hyperGrid}{data.frame or list of lists of hyper-parameters to try}

\item{rounds}{outer rounds}

\item{clientFrac}{fraction of clients sampled per round}

\item{resultDirectory}{directory for logs and results}

\item{epsilon}{convergence tolerance}

\item{mirai}{if TRUE, use \code{mirai}; otherwise use PSOCK workers}
}
\value{
a data.frame of outer-fold test metrics + best hyperparams
}
\description{
Outer + inner CV for federated hyperparameter tuning
}
