% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep.R
\name{filterCovariateRef}
\alias{filterCovariateRef}
\title{Filter covariate references for a named experiment feature set}
\usage{
filterCovariateRef(
  covariateRef,
  featureSet = NULL,
  covariateIds = NULL,
  analysisIds = NULL
)
}
\arguments{
\item{covariateRef}{covariate reference data.frame}

\item{featureSet}{one of "all", "ageSex", "phenotypes", "ageSexPhenotypes"}

\item{covariateIds}{optional explicit covariate ids to retain}

\item{analysisIds}{optional explicit analysis ids to retain}
}
\value{
filtered covariateRef data.frame
}
\description{
Filter covariate references for a named experiment feature set
}
