% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep.R
\name{getClientFeatures}
\alias{getClientFeatures}
\title{Get the covariateRef data.frame from a plpData object}
\usage{
getClientFeatures(plpData)
}
\arguments{
\item{plpData}{a plpData object}
}
\value{
covariateRef data.frame with covarateId, covariateName, etc.
}
\description{
Get the covariateRef data.frame from a plpData object
}
