% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MathUtils.R
\name{gradLogistic}
\alias{gradLogistic}
\title{Stochastic gradient of logistic loss}
\usage{
gradLogistic(weights, xMatrix, yLabels)
}
\arguments{
\item{weights}{numeric vector of parameters}

\item{xMatrix}{numeric matrix (n by p)}

\item{yLabels}{numeric vector of binary labels}
}
\value{
numeric vector (length p)
}
\description{
Stochastic gradient of logistic loss
}
