% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPrep.R
\name{loadClientData}
\alias{loadClientData}
\title{Load one client's PLP data and build feature/label matrices}
\usage{
loadClientData(path, popSettings)
}
\arguments{
\item{path}{path to a client's folder}

\item{popSettings}{populationSettings for PatientLevelPrediction}
}
\value{
plpData object with population
}
\description{
Load one client's PLP data and build feature/label matrices
}
