% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MathUtils.R
\name{proxL1}
\alias{proxL1}
\title{Soft-threshold l1-proximal operator}
\usage{
proxL1(z, alphaLambda, intercept = TRUE)
}
\arguments{
\item{z}{numeric vector (dual state)}

\item{alphaLambda}{nonnegative scalar (alpha * lambda)}

\item{intercept}{wether there is an intercept}
}
\value{
numeric vector (primal state)
}
\description{
Soft-threshold l1-proximal operator
}
