% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{serverInitDualAveragingCpp}
\alias{serverInitDualAveragingCpp}
\title{Initialize the C++ dual-averaging server state}
\usage{
serverInitDualAveragingCpp(config)
}
\arguments{
\item{config}{algorithm configuration list}
}
\value{
server state list
}
\description{
Initialize the C++ dual-averaging server state
}
