% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FederatedDualAveraging.R
\name{serverRoundDA}
\alias{serverRoundDA}
\title{One server round}
\usage{
serverRoundDA(serverState, clientReports, config)
}
\arguments{
\item{serverState}{state of server}

\item{clientReports}{what to get from clients}

\item{config}{configuration list}
}
\description{
One server round
}
