% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{serverRoundDualAveragingCpp}
\alias{serverRoundDualAveragingCpp}
\title{Run a C++ dual-averaging server round}
\usage{
serverRoundDualAveragingCpp(serverState, clientReports, config)
}
\arguments{
\item{serverState}{current server state}

\item{clientReports}{list of client update reports}

\item{config}{algorithm configuration list}
}
\value{
list with updated server state and report
}
\description{
Run a C++ dual-averaging server round
}
