% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Client.R
\name{clientInit}
\alias{clientInit}
\title{Initialize a client (not strictly needed for Euclidean mirror)}
\usage{
clientInit(w0)
}
\arguments{
\item{w0}{initial primal (p‐vector)}
}
\value{
named list with z = ∇h(w0) (for Euclid h=½∥w∥² this is
z=w0)
}
\description{
Initialize a client (not strictly needed for Euclidean mirror)
}
