% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FederatedCv.R
\name{federatedNestedCv}
\alias{federatedNestedCv}
\title{Outer + inner CV for federated hyperparameter tuning}
\usage{
federatedNestedCv(
  clientHosts,
  clientPaths,
  popSettings,
  algorithm,
  hyperGrid,
  rounds,
  clientFrac,
  epsilon = 1e-06
)
}
\arguments{
\item{popSettings}{(only if you need to re‐build population)}

\item{hyperGrid}{data.frame or list of lists of hyper‐parameters to try}

\item{rounds}{outer rounds}

\item{dataList}{list of length M, each element $X,$y}

\item{w0}{initial parameter vector}

\item{k}{clientFrac}
}
\value{
a data.frame of outer‐fold test metrics + best hyperparams
}
\description{
Outer + inner CV for federated hyperparameter tuning
}
