% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orchestrator.R
\name{fitFederated}
\alias{fitFederated}
\title{Run any registered federated learning algorithm end-to-end}
\usage{
fitFederated(cl, algorithm, config, verbose = TRUE)
}
\arguments{
\item{cl}{cluster object}

\item{algorithm}{name, e.g. "DualAvg", "FedAvg"}

\item{config}{algorithm configuration list}

\item{verbose}{if TRUE, print optimization progress}
}
\value{
w
}
\description{
Run any registered federated learning algorithm end-to-end
}
