% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orchestrator.R
\name{fitFederated}
\alias{fitFederated}
\title{Run \emph{any} registered FL algorithm end‐to‐end}
\usage{
fitFederated(cl, algorithm, config, verbose = TRUE)
}
\arguments{
\item{cl}{cluster object}

\item{algorithm}{name, e.g. "DualAvg", "FedAvg"}

\item{config}{list(etaClient, etaServer, K, rounds, lambda, clientFrac)}
}
\value{
w
}
\description{
Run \emph{any} registered FL algorithm end‐to‐end
}
